Update AlgorithmX.py
Removed the August 13th changes and updated comments. Specifically removed the functionality that allowed rows to be preselected before running Algorithm X. The only option allowed now is to restrict the rows so that Algorithm X has no choice but to select rows that must be part of the solution, such as prefilled cells in a logic puzzle.